+++ internal/acceptance_tests/blocks/logging_heroku_nested.tf
@@ -1,0 +1,7 @@
+logging_heroku {
+  name = "{{.LOGGING_HEROKU_NAME}}"
+  url  = "https://example.herokuapp.com/log"
+  authentication = {
+    token = "test-heroku-token"
+  }
+}