+++ internal/acceptance_tests/blocks/logging_heroku_nested_multi.tf
@@ -1,0 +1,15 @@
+logging_heroku {
+  name = "{{.LOGGING_HEROKU_NAME_1}}"
+  url  = "https://example.herokuapp.com/log"
+  authentication = {
+    token = "test-heroku-token"
+  }
+}
+
+logging_heroku {
+  name = "{{.LOGGING_HEROKU_NAME_2}}"
+  url  = "https://example.herokuapp.com/log"
+  authentication = {
+    token = "test-heroku-token"
+  }
+}